requirements.txt
python-pptx>=0.6.21
openpyxl>=3.1.2
Pillow>=10.0.0
anthropic>=0.39.0
click>=8.1.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
fpdf2>=2.7.9
itsdangerous>=2.2.0
